+++ YouTube-client-app/src/app/youtube/pages/main-page/cards-list/card-item/card-item.component.scss
