+++ YouTube-client-app/src/app/youtube/pages/main-page/cards-list/card-item/card-item.component.ts
@@ -1,9 +1,9 @@
 import { CommonModule } from '@angular/common';
 import { Component, Input } from '@angular/core';
 
-import { ButtonComponent } from '../../../../shared/button/button.component';
-import { CardColorDirective } from '../../../directives/card-color.directive';
-import { CardItemModel } from '../../../models/card-item.model';
+import { ButtonComponent } from '../../../../../shared/components/button/button.component';
+import { CardColorDirective } from '../../../../directives/card-color.directive';
+import { CardItemModel } from '../../../../models/card-item.model';
 
 @Component({
   selector: 'app-card-item',
