+++ YouTube-client-app/src/app/youtube/pages/main-page/cards-list/cards-list.component.scss
